Use same protocol for thrift clients and servers (#1318)
Problem

Right now it's possible to configure thrift clients and servers to use
different `thriftProtocol` settings, but linkerd does not convert the
thrift payloads when forwarding requests and responses, so this
configuration is invalid. It's also causing problems for the serializer
used when `thriftMethodInDst` is set to true.

Solution

Move `thriftProtocol` to be a router param, which ensures that both
clients and servers use the same protocol.