Add HTTP/2 trace initializer and tracing code (#732)
Problem

h2 routers did not include a trace initializer, so spans were not being
assigned hierarchical trace ids nor protocol-specific trace annotations.

Solution

Add H2TraceInitializer, as well as an h2-specific TracingFilter. I also
refactored the HttpTraceInitializer into an abstract class that could be
shared by both the http and h2 initializers. And I fixed a trace
annotation issue that arose due to a change in the Label param on the
path stack; that stack now includes a RouterLabel param, which we use
for the RPC annotation on server spans.