Cache dtab observations in consul dtab store (#1334)
Each call to ConsulDtabStore.observe creates a new observation, even if there
is already an observation active for that namespace.  This means that namerd
will have N identical watches on a dtab namespace where N is the number of
names observed in that namespace.

Add a cache of activities so that there is only one activity per namespace.
These activities are reference counted and only kept open if observed so there
is little cost to caching them forever.

Manually verified that establishing multiple dtab watches on the same namespace
via the http or grpc interface only results in a single connection to consul.