Fix some minor dev issues (#1376)
Problem

The Linkerd source code tarball won't build readily on new systems due
to our build files relying on the build being run from within a git
checkout, and our `protoc` script depending on the `unzip` utility.

Solution

If we're not in a git checkout, assume that we are building a stable
version. And switch from `unzip` to `jar xf` for extracting zip files.