Accept named targetPort
We assume that the targetPort property of a service port is an int but the k8s
spec allows it to be either an int or a string.  If the targetPort is a string
(named port) then linkerd throws a ClassCastException.

We parse targetPort as a string and then attempt to convert it to an int.  If
this conversion is successful we treat targetPort as a port number otherwise
we treat it as a port name.

Added a test in EndpointsNamerTest.