Istio finagle upgrade (#1437)
* Move CircleCI parallelism setting (#1400)

* Update to finagle 6.45.0, netty 4.1.10 (#1389)

* Update to finagle 6.45.0, netty 4.1.10

* Fix com.twitter.common.util version conflict

* Pin the base docker image's JRE version (#1422)

Problem

By depending on the "java:openjdk-8-jre" tag, docker just uses whatever
minor version of the JRE is available the first time that image is
pulled. This can lead to unpredictable minor versions across releases.
For instance:

```
$ docker run --rm --entrypoint=sh buoyantio/linkerd:0.9.1 -c 'echo $JAVA_DEBIAN_VERSION'
8u111-b14-2~bpo8+1
$ docker run --rm --entrypoint=sh buoyantio/linkerd:1.0.0 -c 'echo $JAVA_DEBIAN_VERSION'
8u102-b14.1-1~bpo8+1
$ docker run --rm --entrypoint=sh buoyantio/linkerd:1.0.2 -c 'echo $JAVA_DEBIAN_VERSION'
8u111-b14-2~bpo8+1
```

Solution

Pin our base image to a known recent minor version: "openjdk:8u131-jre".
As part of this change I'm also switching us from the "java" image to
the "openjdk" image, since "java" is deprecated.

* Remove linkerd-acceptance test files (#1424)

Problem

The linkerd-acceptance.sh test script is flaky and was disabled #1226.

Solution

The script is still flaky even after the finagle upgrade, so we're just
going to remove it entirely.

* Modify Istio components to compile with new finagle version