Add jitter to Marathon namer poll interval (#171) (#1448)
* Add jitter to Marathon poll attempts (#171)

For very large clusters, if a bunch of linkerds all poll the Marathon namer at exactly the same time, the namer
can be overloaded.

I've added a slight random jitter to poll attempts, based on the formula suggested by @adelong in comments
on PR #772. The jitter amount is configurable in YAML. Note that I decided to just pass a function from
MarathonInitializer to AppIdNamer, rather than pass a Stream as @adelong recommended on #772 – this
seemed simpler & thus more elegant. Annotating the jitter function with `@inline` should prevent it from
creating a new call stack, so performance is probably about the same as using a `Stream`.

I've added a handful of tests for this: we test that the correct values are parsed from the config file,
that a respectable sample of the jittered TTLs are within the expected range, and that at least 50% of
the random values are unique.

Fixes #171

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Fix incorrect lower bound in jitter tests

I forgot to subtract the maximum jitter amount from the expected lower bound - my bad!

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Fix incorrect lower bound in jitter tests

I forgot to subtract the maximum jitter amount from the expected lower bound - my bad!

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Make jitter tests expect a more reasonable number of collisions

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Reduce the number of expected unique TTLs again

@adelong suggested this
(https://github.com/linkerd/linkerd/pull/1448#discussion_r124324987)
and...he's probably right.