Add failing test for issue #1439
This commit adds a test that reproduces #1439, by simulating 20
consecutive failures and asserting that the stream of durations produced
by the backoff policy are constant. This test currently fails. This
commit also adds a test that asserts that the default backoff policy is
used when no backoff policy is specified in YAML, to ensure fixes for
this issue don't break that behaviour.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>