Decide to use default failure accrual backoff lazily (#1439)
Problem:
As described in issue #1439, custom failure accrual backoff policies
specified in YAML are currently ignored – linkerd uses the default
backoff policy regardless of what is specified.

Solution:
I traced the issue to the `backoffOrDefault` value member in
`FailureAccrualInitializer`. This value, which is set to the class's
`backoff` member if it is defined, and to the default otherwise, was
being evaluated eagerly when instances of the initializer are
constructed. However, the `backoff` member, which is mutable, is
*always* set to `None` initially, and is mutated to the config value
*after* the class is instantiated. Thus, `backoffOrDefault` is always
set to the default.

I've fixed the issue simply by changing `backoffOrDefault` to a lazily-
evaluated value.

Validation:
In an attempt to localize this fault, I've added a number of unit tests.

I initially thought the fault was in how we parse configs for failure
accrual policies, so I've added several tests for config parsing. I then
proceeded to add tests which simulate failure accrual by repeatedly
calling the failure method on a failure accrual policy and asserting the
resulting Stream of values contains the correct probe durations.

Some of my unit tests were written with the ScalaCheck property testing
library, so our tests now depend on ScalaCheck.

Fixes #1439

Signed-off-by: Eliza Weisman <eliza@buoyant.io>