Decide to use default failure accrual backoff lazily (#1439) (#1451)
* Add tests for parsing failure accrual configs

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add a dependency on ScalaCheck for property-based testing.

I based this on the way @stevej added ScalaCheck in 542d6bde9ff3dc6ed3ce8da6d6098e60fa9ff435,
so hopefully there'll be fewer merge conflicts.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add tests for backoff configs producing values within the expected range

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add test for uniqueness of jittered backoff durations

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add failing test for issue #1439

This commit adds a test that reproduces #1439, by simulating 20
consecutive failures and asserting that the stream of durations produced
by the backoff policy are constant. This test currently fails. This
commit also adds a test that asserts that the default backoff policy is
used when no backoff policy is specified in YAML, to ensure fixes for
this issue don't break that behaviour.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Decide to use default failure accrual backoff lazily (#1439)

Problem:
As described in issue #1439, custom failure accrual backoff policies
specified in YAML are currently ignored – linkerd uses the default
backoff policy regardless of what is specified.

Solution:
I traced the issue to the `backoffOrDefault` value member in
`FailureAccrualInitializer`. This value, which is set to the class's
`backoff` member if it is defined, and to the default otherwise, was
being evaluated eagerly when instances of the initializer are
constructed. However, the `backoff` member, which is mutable, is
*always* set to `None` initially, and is mutated to the config value
*after* the class is instantiated. Thus, `backoffOrDefault` is always
set to the default.

I've fixed the issue simply by changing `backoffOrDefault` to a lazily-
evaluated value.

Validation:
In an attempt to localize this fault, I've added a number of unit tests.

I initially thought the fault was in how we parse configs for failure
accrual policies, so I've added several tests for config parsing. I then
proceeded to add tests which simulate failure accrual by repeatedly
calling the failure method on a failure accrual policy and asserting the
resulting Stream of values contains the correct probe durations.

Some of my unit tests were written with the ScalaCheck property testing
library, so our tests now depend on ScalaCheck.

Fixes #1439

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Remove JUnit runner, unused imports from ConfigTest

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add missing assertions in ConfigTest

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Use assert rather than matchers in ConfigTest

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Allow jittered backoff min/max to be equal.

Apparently Finagle allows this. Huh.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Remove `yield outcomeOf` that wasm causing spurious test results

* Remove vestigial for-expression in ConfigTest

* Ensure `probeDelays` contains _only_ 4 Nones

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Assert subsequent probe delays are defined in jittered backoff test

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Add test for consecutive failure policies with any number of failures

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Minor code style fixes

* Change `backoffOrDefault` from `lazy val` to `def`

As @adelong requested in
https://github.com/linkerd/linkerd/pull/1451#discussion_r124867291

* Replace Table in ConfigTest with Seq

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* Remove unneeded imports

Signed-off-by: Eliza Weisman <eliza@buoyant.io>