Fix null pointer in delegator when address is not resolable
If the delegator gets an unresolvable address, it throws a
NullPointerException.

Check that the address is not null before using it in a delegation.

I wasn't able to reproduce this issue because unresolable addresses appear as
Negs for me.  It's unclear to me how to get an Addr.Bound with an unresolvable
address in it.  However, based on the stack trace in the bug report, this
should fix the problem.