Add HttpResponseException
There are several places in Linkerd where it is desirable to stop processing
the current request and return a specific response.  However, many of those
places (such as identifiers, for example) do not expose any interface for
returning responses.

We introduce the HttpResponseException which contains a Response object.  By
returning a failed Future with this exception, we can halt processing of the
current request and return the given response.