Add http identifier endpoint to linkerd
Problem:
As identifiers increase in complexity (see IstioIngressIdentifier), it becomes more
difficult to debug them.

Solution:
Adds a config admin option `httpIdentifierPort` that if set serves a json endpoint
on that port. Given a request, the endpoint returns a map where the keys are
router labels and the values are identifier responses configured for those routers.