Add interpreter for reading from a k8s ConfigMap
This commit adds an interpreter for reading Dtabs from a Kubernetes ConfigMap – as described in issue #1506.

I've built on @klingerf's work in 9a7c55f1449dc44fa95a31ebf74347c1fb9713ad and my changes to `Watchable` e8985e2f1a66ec6726216d5927510dc7b09e14e9 to refactor Kevin's old interpreter code. This interpreter uses the new `Watchable` API for Kubernetes object resources to establish a watch on the requested ConfigMap, access a key containing a dtab, and use that dtab to resolve routes. Updates to the dtab are reflected in the interpreter. The new interpreter is currently marked as experimental until it can be thoroughly tested in production.

I've added unit tests for getting ConfigMaps to ApiTest and new unit tests for the `ConfigMapInterpreterInitializer` configuration. Additionally, I believe that Kevin has tested this a bit in a live Kubernetes cluster. I've also added documentation on using the new interpreter to `linkerd/docs/interpreters.md`.

Closes #1506