Add stream classification & advanced observability for H2 (#1517)
As described in #1508, I've added a `H2StreamClassifier` type. This type is equivalent to a `ResponseClassifier` for H2, except that it takes an `Option[Try[Frame]]` containing the final frame in the response stream, as well as the request and response. I've replaced uses of `ResponseClassifier` in the H2 package with `H2StreamClassifier`, and modified`StreamStatsFilter` to accept a `H2StreamClassifier` as a stack param. 

In order to pass the final frame to the `H2StreamClassifier` without violating flow control, I've added a `StreamProxy` class which allows an `onFrame: Try[Frame] => Unit` function to be attached to an H2 stream. `StreamStatsFilter` uses the `onFrame` function to check whether each frame is the final frame in the stream, and if so, passes it to the stream classifier. I've also used the `onFrame` function to add the following per-frame metrics:

+ `response/stream/total_bytes`: the total size (in bytes) of all the data frames in the response stream 
+ `request/stream//total_bytes`: tthe total size (in bytes) of all the data frames in the request stream 

I've added unit tests for the new functionality to `StreamStatsFilterTest`.