Add ExceptionStatsHandler to StreamStatsFilters (#1543)
While working on tests for #1542, I noticed that the standard `StatsFilter` also has an `ExceptionStatsHandler` stack param, which `StreamStatsFilter` lacks. This param is used for counting the number of occurrences of specific exceptions that occur.

I thought we might want to have that for stream stats as well (and @adleong confirmed). This PR adds the `ExceptionStatsHandler` to both `StreamStatsFilter` and the `LocalClassifier` and `PerDstPath` stream stats filter derivatives added in #1542. 

The only aspect of this PR that I think is complicated enough to comment on in detail is that I've added a public re-export of `ResponseClassificationSyntheticException` to `com.twitter.finagle.buoyant`. This exception appears to be used by the `ExceptionHandler` for cases where the classifier determined that there was a failure, but there was no associated exception thrown. Unfortunately, it's declared as `private[finagle]`, so in order to use it in `StreamStatsFilter` (which lives in the `io.buoyant.router.h2` package), I needed to re-export it from somewhere that lives inside the `finagle` package.

I've re-enabled the tests in `PerDstPathStreamStatsFilterTest` that expect per-exception stats counters. These were commented out when I modified that test class from the non-stream-aware version to test #1542, since we weren't currently generating those stats. Now that we are, they can be re-enabled.