Allow early H2 classification (#1545)
The H2StreamClassifier type requires that a response stream be complete before
the response can be classified.  This is so that classifiers may use
information from the Trailers, such as gRPC status code, to inform
classification.  However, for classifiers that only use information from the
response Headers, classification is unnecessary delayed until the stream is
complete.  Furthermore, once retries are introduced, responses will be held
and buffered until the reponse can be classified or the buffer is overfull.
Holding the response until the response stream completes introduces an
unnecessary delay when enough information exists in the response Headers to
classify that response.

Rename H2StreamClassifer to H2Classifer and enrich it to include two partial
functions: responseClassifier and streamClassifier.  When the response future
completes, the responseClassifier should be consulted.  If the
responseClassifier partial function is not defined for this input, the
streamClassifier partial function should be consulted once the response stream
is complete.  This allows classifiers to classify "early" when there is enough
information in the response headers to make a classification.