H2 retries (#1523)
Finagle's retry mechanisms do not work for H2 because of the presence of request
and response streams. When Linkerd is processing a response from a downstream
service, it may not have enough information to determine if the request was
retryable until the response stream completes. Furthermore, the request stream
has already been (or is already being) streamed to the first downstream endpoint
so we cannot send that same stream to a second downstream endpoint in the case
of a retry.

We add a stream aware ClassifiedRetriesFilter which uses BufferedStream and the
stream aware H2Classifiers. When a request comes in, the request stream is
buffered so that it can be replayed to a second downstream endpoint if a retry
is necessary. When the response comes back, the response stream is buffered as
well and no response object is returned until we can make a determination about
retryability or we give up. Depending on the H2Classifier, we may be able to
determine the retryability immediately (if the H2Classifier's responseClassifier
is defined) or we may need to wait until the response stream is complete and use
the H2Classifier's streamClassifier. If the response buffer becomes full or if a
timeout is exceeded, we give up and return the current response stream from the
buffer. If we determine that a retry should be made, we fork a new request
stream from the request stream buffer and send it to a downstream endpoint.