Use request local classifier for failure accrual
FailureAccrualFactory uses the response classifier configured on the stack
instead of the response classifier from the request local context.  This means
that circuit breakers will trip on different success criteria than is used for
stats reporting.

Replace Finagle's FailureAccrualFactory with a subclass which uses the response
classifier from the request local context.