Add H2StreamClassifier to ClassifierFilter (#1544)
Previously, the server stack `ClassifierFilter` did not use a stream-aware `H2StreamClassifier`, and wasn't able to set the `l5d-success-class` header in the trailers frame of H2 streams. The `ClassifierFilter` needs to be able to make classification decisions based on the final frame of a `Stream` as well as the status code, and if the final frame is a Trailers frame, the response class header should be set into the trailers frame.

In this PR, I've rewritten the H2 `ClassifierFilter` to use `H2StreamClassifier` as a stack parameter, and to write response class headers into Trailers frames, if they are present. Otherwise, the header is written to the response headers.

I've modified `StreamClassificationEndToEndTest` to include several e2e tests for these changes.

Closes #1538