Use correct StatsFilter and Classifier for H2 (#1558)
Due to a discrepncy in the LocalClassifierStatsFilter role, the H2 client stack
did not include the LocalClassifierStreamStatsFilter.  This was causing the
success/failures stats to be incorrect at the client level.  Furthermore, the
PerDstPathStreamStatsFilter was reading the H2Classifier from the local context
outside of the request path which was causing the wrong classifier to be used.
This was causing the client/*/service success/failure stats to be incorrect.

Fix the LocalClassifierStatsFilter role so that
LocalClassifierStreamStatsFilter gets added to the H2 client stack.  Modify
PerDstPathStreamStatsFilter to read the H2Classifier in the request path.