Grpc-Status code may come in H2 headers (#1567)
The grpc response classifiers assume that the grpc-status code will appear in
a trailers frame but this is not always the case.  If a response has no data
frames, the grpc-status may appear in the headers frame and the stream will be
empty.  The grpc response classifiers fail to correctly classify in this case.

Change the grpc response classifiers to check the headers frame for a
grpc-status code as part of early classification.