Ensure headers-only streams are closed (#1569)
When Netty4StreamTransport writes a message with an empty stream, it will
attempt to read frames off of the stream until it is empty before closing the
local side of the stream.  Since the stream is empty, it will never close
local and the stream will persist in the streams hashtable forever causing
a memory leak.

Do not attempt to read frames off of an empty stream, instead closing local
immediately after headers have been written.