Correct handling of non-existent resources
The test suite for ServiceNamer dictates that an unknown service name
should return a negative NameTree, BUT watch the service name in case
it is created later. In order to fix this, I modified Watchable.get
to return None if the requested path returns 404. Watchable.activity
still starts the watch if None is returned. This required some
modification to other existing classes.