Ignore Reader.ReaderDiscarded exception (#1609)
As seen in #1594 – Linkerd occasionally throws `ReaderDiscarded` exceptions when discarding a response on retries. Since nothing is actually wrong, we want to silence these exceptions.

I've added a `Try` to the call to `response.discard` in `RetryFilter.scala`. The `IOException` "Connection reset by peer" is not silenced, since it appears to be both thrown and caught in code we don't own & it would be difficult to quash this exception only on retries.

Tested this against the Kubernetes config described by @esbie in https://github.com/linkerd/linkerd/issues/1594#issue-251047691; the `RetryDiscarded` exception no longer appears in the logs. 

Closes #1594.