Kubernetes API refactor (#1603)
This commit contains three major features: first, a refactor of the `Watchable` trait in `io.buoyant.k8s`; second, a new interpreter for reading dtabs from a Kubernetes ConfigMap; and third, refactoring of the Kubernetes `EndpointsNamer` and `ServiceNamer`s to watch individual Kubernetes API resources rather than the entire namespace.

### Make Kubernetes Objects watchable (#1527)

Previously, only Kubernetes list resources were watchable, since our code for establishing watches on Kubernetes resources used the `?watch=true` path param.

I've rewritten our code to establish watches using the `$API/$NS/watch/$RESOURCE` endpoint, instead, thus making the `KubeObject` type watchable as well as `KubeList`. I've also factored out some commonly repeated code for creating an `Activity` to watch a Kubernetes resource, and put that in the `Watchable` trait.

The unit tests in`EndpointsNamerTest`, `ServiceNamerTest`, and `ApiTest` have all been updated to cover the new `Watchable` implementation. Additionally, I believe @klingerf has tested this somewhat on a live Kubernetes cluster.

### Add interpreter for reading from a k8s ConfigMap (#1532)
This commit adds an interpreter for reading Dtabs from a Kubernetes ConfigMap – as described in issue #1506.

I've built on @klingerf's work in 9a7c55f and my changes to `Watchable` e8985e2 to refactor Kevin's old interpreter code. This interpreter uses the new `Watchable` API for Kubernetes object resources to establish a watch on the requested ConfigMap, access a key containing a dtab, and use that dtab to resolve routes. Updates to the dtab are reflected in the interpreter. The new interpreter is currently marked as experimental until it can be thoroughly tested in production.

I've added unit tests for getting ConfigMaps to ApiTest and new unit tests for the `ConfigMapInterpreterInitializer` configuration. Additionally, I believe that Kevin has tested this a bit in a live Kubernetes cluster. I've also added documentation on using the new interpreter to `linkerd/docs/interpreters.md`.

### K8s namers watch individual objects, rather than the entire ns (#1587)
Right now, the `io.l5d.k8s` and `io.l5d.k8s.ns` namers setup a watch on the List endpoints API, which streams update events for all of the endpoints in a given namespace. This can potentially lead to receiving a lot of events for endpoints that Linkerd is not trying to actually route to. It can also lead to a memory leak, in the case that Linkerd is watching a namespace with very frequent updates (e.g. `kube-system`/ #1361).

My PR #1527 added the ability to watch individual Kubernetes API objects, rather than watching every object in a namespace. This pull request rewrites the `EndpointsNamer` and `ServiceNamer`s to use the new watch code added in #1527, both for the `/namespaces/{ns}/endpoints/{service}` API endpoint and for the `/namespaces/{ns}/services/{service}` endpoint that we use for resolving numbered port mappings.

The `EndpointsNamer` caches watches on each (namespace, service name) combination and reuses existing watches so that we don't create a whole bunch of threads all watching the same API response. Furthermore, the watch on `.../services/...` is only established if it is needed (i.e. the namer was given a numbered rather than named port). I've also attempted to simplify the caches used in `EndpointsNamer` as much as possible.

Note that I've also made a change to the `Watchable.get()` method signature; it now returns a `Future[Option[G]]` rather than `Future[G]`. The test suite for` ServiceNamer` dictates that an unknown service name should return a negative `NameTree`, _but_ watch the service name in case it is created later. In order to fix this, I modified `Watchable.get()`to return `None` if the requested path returns 404. `Watchable.activity()` still starts the watch if `None` is returned. This required some modification to other existing classes, but appears to not have broken anything.

I've updated the existing unit tests in `EndpointsNamerTest`and `ServiceNamerTest` and confirmed they all work in dev. I've also modified other Kubernetes API tests to reflect the change to `Watchable.get()` I mentioned above.

Closes #1534
Closes #1575
Closes #1361