Actually fix my awful broken code
I finally checked out this branch in IntelliJ instead of making edits on
GitHub. Sorry for all the broken builds.

This has the side benefit of making forks have the type
`BufferedStream.Fork` which I think will make a lot more sense in the
debugger than `BufferedStream$anon` (which I assumed was a fn rather
than a class, at first).