Add option to filter consul nodes by health statuses (#1608)
Adding the healthStatuses parameter to linkerd consul api #1601.

The only concern I have is the way the `healthStatuses` are used to filter the nodes are passed to `HealthApi`, I could move the parameter into the serviceNodes method but that would require also modifying `CatalogueApi` and a number of intermediate calls. There may be another approach that I am missing though.

A few other things are worth highlighting:

- Each consul node can have more than one health check associated with it, for this reason the checks for an individual node are aggregated into a single health status before the filtering is applied. This is the same approach as taken in [consul template](https://github.com/hashicorp/consul/blob/master/api/health.go) (in particular see line 42.)

- There is some conflict with the existing `useHealthChecks` parameter, for backwards compatibility my implementation currently makes `useHealthChecks = True` override the healthStatuses parameter to be just `passing`. There may be a better approach to this, such as throwing an exception if both are defined or deprecating `useHealthChecks` in favour of `healthStatuses`.

- If `healthStatuses` is passed as just `passing` (or if `useHealthChecks = True`) then we could 
still use the `passing` parameter of the consul api to filter on the server side. However this would complicate the code for potentially little/no real gain.

- While `healthStatuses` takes an array of statuses it also allows any `any` parameter for convenience, which is also the default parameter. If no value is provided to `healthStatuses` then
it defaults to `any` which is equivalent to the existing behaviour.

Closes #1601