Handle K8s Delete Events Correctly (#1622)
When Linkerd receives a DELETE event from the Kubernetes API, it treats the
contents of the event as the new state.  However, the semantics of the API are
that the contents of the event are the previous state and the resource should
be deleted entirely.  This causes Linkerd to get stuck in a bad state when
watching a resource as it is being deleted.

Treat DELETE events as total deletions and ignore the contents of the event.

Manually tested on Minikube.