Fix K8s namers not updating cached name Vars (#1638)
The way `stabilize()` is currently implemented means that when a service is scaled down and then scaled back up, or removed and then re-created, the namer returns a new `Var` with the new address of that service. However, when a client stack for the service is created, it's given the original `Var` and cached, so it doesn't receive the updated address.

I've changed `stabilize()` to update the previous `Var`, rather than creating a new one.

I've added tests in `EndpointsNamerTest` based on events fired by Kubernetes while I was trying to debug this issue – there's a test that addresses are correctly updated after a scale down/scale up, and a test that references to the original `Var` are now properly updated.

Fixes #1626.
Fixes #1635.
Fixes #1636.