Remove watch on service list endpoint from Consul namer (#1646)
* Add test for CatalogAPI handling empty response
As reported [on Discourse](https://discourse.linkerd.io/t/linkerd-cpu-usage-climbs-over-time/274), users running Linkerd with the Consul namer have observed a steady increase of CPU usage. The current implementation of the Consul namer establishes a watch on the `/catalog/services` Consul endpoint, which returns a list of all services in the Consul datacenter. This means that Linkerd will always be notified of changes to the Consul services list, regardless of whether or not it cares about those services. When there are a great deal of services in Consul, tracking updates to this list becomes very demanding. 

We only use the services list endpoint to determine the existence of a service. However, the Consul API endpoint to look up the list of nodes for a service will return an empty list if the service does not exist, which can also be used to determine the existence of a service. I've simplified the Consul namer to use only the `/catalog/service/:service` endpoint instead of the `/catalog/services` endpoint. This means that we only watch the services we care about and are not notified on all changes. Furthermore, this allows the Consul namer to be made much simpler – the entire `DcServices` class can now be removed.

I've reproduced the performance issue by using nginx to simulate Consul. I configured nginx to statically serve a service list JSON file with 1000 services with 10 tags each. Nginx will always respond instantly, which is equivalent to watch events constantly firing. A single HTTP request to Linkerd 1.2.1 causes it to open a watch on this service list, sending CPU usage to hit 100% almost immediately. Against my branch, this no longer happens.

I had to make some changes to the unit tests in `ConsulNamerTest.scala` to reflect the new behavior of the namer. Of note is that we no longer expect the `dc/` set of stats to be incremented, as these stats were related to the watch on the services list endpoint, which no longer happens.
* Factor out `stabilize()` from K8s for reuse in Consul

* Rough first pass on individual svc lookup

* "Namer stays pending" test now has serviceNodes endpoint

* Fix lookup activity not pending

* Remove dc stats that are no longer incremented from tests

* Update blocking call test to block on serviceNodes

* Fix test for failures on initial GET request

* Fix test for pending when DC is nonexistant

* Correct test description

* Remove dead code in tests

* Remove dead code

* Review feedback

* Revert accidentally-committed changes to Consul example